rename config to settings
