+++ dance/__init__.py
@@ -1,11 +1,11 @@
 import logging
 
-from dance import config
+from dance import settings
 
 logger = logging.getLogger("dance")
 
 __version__ = "1.0.1-dev"
 __all__ = [
-    "config",
     "logger",
+    "settings",
 ]
